- NCubeAppContext bean removed from beans.xml and declared in NCubeConfiguration - ReflectiveProxy now takes an NCubeMutableClient as a constructor argument (NCubeManager is an NCubeMutableClient), and this is bound at bean creation time. - NCubeRuntime constructor now defines what the bean name of the target will based on whether the CallableBean is an instance of JsonHttpProxy or not.  This is important because Spring must locate the controller by name on another server.  In this case, the name of the Bean to locate is the NCubeController.
When running on the combined server, the beanName for NCubeRuntime is feed tp to the CallableBean (ReflectiveProxy), but the reflective proxy targets who it was bound to at construction time (NCubeManager).  So the beanName in this case is used only as part of a key for the method cache inside of the ReflectiveProxy.